+++ src/bin/bench.rs
@@ -31,6 +31,10 @@ struct Args {
     #[clap(long, default_value_t = 1)]
     concurrency: usize,
 
+    /// Enable warmup
+    #[clap(long, default_value_t = false)]
+    warmup: bool,
+
     /// Request only http/get
     #[clap(long, default_value_t = false)]
     only_httpget: bool,
@@ -52,14 +56,15 @@ async fn main() -> anyhow::Result<()> {
 
     let count = Arc::new(Mutex::new(args.count));
 
-    let elapsed_jsonrpc = Arc::new(Mutex::new(Duration::ZERO));
-    let elapsed_httpget = Arc::new(Mutex::new(Duration::ZERO));
+    let elapsed_jsonrpc = Arc::new(Mutex::new((Duration::ZERO, 0)));
+    let elapsed_httpget = Arc::new(Mutex::new((Duration::ZERO, 0)));
 
     try_join_all((0..args.concurrency).map(|_| {
         make_requests(
             url.clone(),
             range.clone(),
             Arc::clone(&count),
+            args.warmup,
             args.only_httpget,
             Arc::clone(&elapsed_jsonrpc),
             Arc::clone(&elapsed_httpget),
@@ -67,15 +72,17 @@ async fn main() -> anyhow::Result<()> {
     }))
     .await?;
 
-    let elapsed_jsonrpc = *elapsed_jsonrpc.lock().await;
+    let (elapsed_jsonrpc, size) = *elapsed_jsonrpc.lock().await;
     println!(
-        "jsonrpc: total {elapsed_jsonrpc:?} / avg: {:?}",
-        elapsed_jsonrpc.div_f64(args.count as f64)
+        "jsonrpc: total {elapsed_jsonrpc:?} / avg: {:?} / total transfered: {}",
+        elapsed_jsonrpc.div_f64(args.count as f64),
+        human_bytes(size)
     );
-    let elapsed_httpget = *elapsed_httpget.lock().await;
+    let (elapsed_httpget, size) = *elapsed_httpget.lock().await;
     println!(
-        "httpget: total {elapsed_httpget:?} / avg: {:?}",
-        elapsed_httpget.div_f64(args.count as f64)
+        "httpget: total {elapsed_httpget:?} / avg: {:?} / total transfered: {}",
+        elapsed_httpget.div_f64(args.count as f64),
+        human_bytes(size)
     );
 
     Ok(())
@@ -85,9 +92,10 @@ async fn make_requests(
     url: Url,
     range: Range<Slot>,
     count: Arc<Mutex<usize>>,
+    warmup: bool,
     only_httpget: bool,
-    elapsed_jsonrpc: Arc<Mutex<Duration>>,
-    elapsed_httpget: Arc<Mutex<Duration>>,
+    elapsed_jsonrpc: Arc<Mutex<(Duration, usize)>>,
+    elapsed_httpget: Arc<Mutex<(Duration, usize)>>,
 ) -> anyhow::Result<()> {
     let mut skip = false;
     loop {
@@ -105,29 +113,37 @@ async fn make_requests(
         let slot = random_range(range.clone());
 
         // warmup
-        if !fetch_slot_get(url.clone(), slot).await? {
+        if warmup && !fetch_slot_get(url.clone(), slot).await?.0 {
             skip = true;
             continue;
-        }
-        if !only_httpget {
-            fetch_slot_json(url.clone(), slot).await?;
         }
 
         // measure
         let ts = Instant::now();
-        let _ = fetch_slot_get(url.clone(), slot).await?;
-        *elapsed_httpget.lock().await += ts.elapsed();
+        let (exists, size) = fetch_slot_get(url.clone(), slot).await?;
+        if !warmup && !exists {
+            skip = true;
+            continue;
+        }
+        let mut locked = elapsed_httpget.lock().await;
+        locked.0 += ts.elapsed();
+        locked.1 += size;
+        drop(locked);
+
         if !only_httpget {
             let ts = Instant::now();
-            fetch_slot_json(url.clone(), slot).await?;
-            *elapsed_jsonrpc.lock().await += ts.elapsed();
+            let size = fetch_slot_json(url.clone(), slot).await?;
+            let mut locked = elapsed_jsonrpc.lock().await;
+            locked.0 += ts.elapsed();
+            locked.1 += size;
+            drop(locked);
         }
     }
 
     Ok(())
 }
 
-async fn fetch_slot_json(url: Url, slot: Slot) -> anyhow::Result<()> {
+async fn fetch_slot_json(url: Url, slot: Slot) -> anyhow::Result<usize> {
     let body = serde_json::json!({
         "jsonrpc": "2.0",
         "method": "getBlock",
@@ -139,7 +155,7 @@ async fn fetch_slot_json(url: Url, slot: Slot) -> anyhow::Result<()> {
     })
     .to_string();
 
-    let response = Client::builder()
+    let mut response = Client::builder()
         .build()
         .context("failed to build http client")?
         .post(url.to_string())
@@ -155,21 +171,25 @@ async fn fetch_slot_json(url: Url, slot: Slot) -> anyhow::Result<()> {
         response.status(),
     );
 
-    let _bytes = response
-        .bytes()
+    let mut size = 0;
+    while let Some(chunk) = response
+        .chunk()
         .await
-        .context("failed to fetch body of jsonrpc request")?;
+        .context("failed to fetch body of jsonrpc request")?
+    {
+        size += chunk.len();
+    }
 
-    Ok(())
+    Ok(size)
 }
 
-async fn fetch_slot_get(mut url: Url, slot: Slot) -> anyhow::Result<bool> {
+async fn fetch_slot_get(mut url: Url, slot: Slot) -> anyhow::Result<(bool, usize)> {
     let slot = slot.to_string();
     if let Ok(mut segments) = url.path_segments_mut() {
         segments.extend(&["block", &slot]);
     }
 
-    let response = Client::builder()
+    let mut response = Client::builder()
         .build()
         .context("failed to build http client")?
         .get(url.to_string())
@@ -178,7 +198,7 @@ async fn fetch_slot_get(mut url: Url, slot: Slot) -> anyhow::Result<bool> {
         .context("failed to send jsonrpc request")?;
 
     if response.status() == StatusCode::BAD_REQUEST {
-        return Ok(false);
+        return Ok((false, 0));
     }
 
     anyhow::ensure!(
@@ -187,10 +207,24 @@ async fn fetch_slot_get(mut url: Url, slot: Slot) -> anyhow::Result<bool> {
         response.status(),
     );
 
-    let _bytes = response
-        .bytes()
+    let mut size = 0;
+    while let Some(chunk) = response
+        .chunk()
         .await
-        .context("failed to fetch body of jsonrpc request")?;
+        .context("failed to fetch body of httpget request")?
+    {
+        size += chunk.len();
+    }
 
-    Ok(true)
+    Ok((true, size))
+}
+
+fn human_bytes(size: usize) -> String {
+    if size > 1024 * 1024 * 1024 {
+        format!("{:.3} GiB", size as f64 / 1024.0 / 1024.0 / 1024.0)
+    } else if size > 1024 * 1024 {
+        format!("{:.3} MiB", size as f64 / 1024.0 / 1024.0)
+    } else {
+        format!("{:.3} KiB", size as f64 / 1024.0)
+    }
 }